users can now be exempted from autocarrot in the config

+++ events/message.js
@@ -30,21 +30,23 @@ module.exports = (client, message) => {
 
 
   if(config.autocarrot.enabled) {
-    const autocarrotWebhook = index.autocarrotWebhook
-    const swearCensors = require("../data/autocarrot/censored_words.json")
-    const swearList = Object.keys(swearCensors)
-
-    let needsCorrecting = false
-    for(i in swearList) {
-      if(message.content.toLowerCase().includes(swearList[i])) {
-        needsCorrecting = true
+    if(!config.autocarrot.exemptedUsers.includes(message.author.id)) {
+      const autocarrotWebhook = index.autocarrotWebhook
+      const swearCensors = require("../data/autocarrot/censored_words.json")
+      const swearList = Object.keys(swearCensors)
+  
+      let needsCorrecting = false
+      for(i in swearList) {
+        if(message.content.toLowerCase().includes(swearList[i])) {
+          needsCorrecting = true
+        }
       }
-    }
-
-    if(needsCorrecting) {
-      autocarrotWebhook(message.author, message.channel, message.content)
-      if(config.autocarrot.deleteOriginalMessage) {
-        message.delete()
+  
+      if(needsCorrecting) {
+        autocarrotWebhook(message.author, message.channel, message.content)
+        if(config.autocarrot.deleteOriginalMessage) {
+          message.delete()
+        }
       }
     }
   }
